feat(client): derive typed HTTP wrappers from shared API
