+++ cardano-mpfs-api/cardano-mpfs-api.cabal
@@ -1,0 +1,47 @@
+cabal-version: 3.4
+name:          cardano-mpfs-api
+version:       0.0.0
+synopsis:      Servant wire contract for the Cardano MPFS API
+description:
+  Lightweight Servant API and JSON wire DTOs shared by the MPFS
+  offchain server and native Haskell clients.
+
+license:       Apache-2.0
+license-file:  LICENSE
+author:        Paolo Veronelli
+maintainer:    paolo.veronelli@gmail.com
+category:      Cardano
+build-type:    Simple
+
+common warnings
+  ghc-options:
+    -Wall -Werror -Wunused-imports -Wunused-packages
+    -Wmissing-export-lists -Wname-shadowing
+    -Wdeferred-out-of-scope-variables -Wpartial-type-signatures
+    -Wredundant-constraints
+
+  default-extensions:
+    DerivingStrategies
+    DuplicateRecordFields
+    OverloadedStrings
+    RecordWildCards
+    StrictData
+
+library
+  import:           warnings
+  default-language: GHC2021
+  hs-source-dirs:   lib
+  build-depends:
+    , aeson              >=2.1  && <2.3
+    , base               >=4.18 && <5
+    , base16-bytestring  >=1.0  && <1.1
+    , bytestring         >=0.11 && <0.13
+    , lens
+    , servant            >=0.20 && <0.21
+    , swagger2
+    , text               >=2.0  && <2.2
+
+  exposed-modules:
+    Cardano.MPFS.API
+    Cardano.MPFS.API.Encoding
+    Cardano.MPFS.API.Types
