+++ cardano-mpfs-client/cardano-mpfs-client.cabal
@@ -34,21 +34,28 @@ library
   default-language: GHC2021
   hs-source-dirs:   lib
   build-depends:
-    , aeson              >=2.1  && <2.3
-    , base               >=4.18 && <5
-    , base16-bytestring  >=1.0  && <1.1
-    , bytestring         >=0.11 && <0.13
-    , cborg              >=0.2  && <0.3
-    , hspec              >=2.11 && <2.12
+    , aeson                >=2.1  && <2.3
+    , base                 >=4.18 && <5
+    , base16-bytestring    >=1.0  && <1.1
+    , bytestring           >=0.11 && <0.13
+    , cardano-mpfs-api
+    , cborg                >=0.2  && <0.3
+    , hspec                >=2.11 && <2.12
+    , http-client          >=0.7  && <0.8
+    , http-types           >=0.12 && <0.13
     , mts:csmt-core
     , mts:csmt-verify
     , mts:mpf-write
-    , operational        >=0.2  && <0.3
-    , text               >=2.0  && <2.2
+    , operational          >=0.2  && <0.3
+    , servant              >=0.20 && <0.21
+    , servant-client       >=0.20 && <0.21
+    , servant-client-core  >=0.20 && <0.21
+    , text                 >=2.0  && <2.2
 
   exposed-modules:
     Cardano.MPFS.Client
     Cardano.MPFS.Client.Bundle
+    Cardano.MPFS.Client.Http
     Cardano.MPFS.Client.Snapshot
     Cardano.MPFS.Client.Verify
     Cardano.MPFS.Client.Verify.DSL
@@ -83,16 +90,21 @@ test-suite unit-tests
     , cardano-mpfs-client
     , cborg                >=0.2  && <0.3
     , hspec                >=2.11 && <2.12
+    , http-client          >=0.7  && <0.8
+    , http-types           >=0.12 && <0.13
     , mts:csmt-core
     , mts:csmt-test-lib
     , mts:csmt-write
     , mts:mpf-test-lib
     , mts:mpf-write
     , text
+    , wai                  >=3.2  && <3.3
+    , warp                 >=3.4  && <3.5
 
   other-modules:
     Cardano.MPFS.Client.BundleSpec
     Cardano.MPFS.Client.Fixtures
+    Cardano.MPFS.Client.HttpSpec
     Cardano.MPFS.Client.SnapshotSpec
     Cardano.MPFS.Client.VerifySpec
