+++ cardano-mpfs-client/test/Main.hs
@@ -1,6 +1,7 @@
 module Main (main) where
 
 import Cardano.MPFS.Client.BundleSpec qualified as BundleSpec
+import Cardano.MPFS.Client.HttpSpec qualified as HttpSpec
 import Cardano.MPFS.Client.SnapshotSpec qualified as SnapshotSpec
 import Cardano.MPFS.Client.VerifySpec qualified as VerifySpec
 import Test.Hspec (hspec)
@@ -10,3 +11,4 @@ main = hspec $ do
     SnapshotSpec.spec
     BundleSpec.spec
     VerifySpec.spec
+    HttpSpec.spec
